backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
firebase-admin>=6.4.0
ecdsa
google-cloud-vision
python-multipart>=0.0.18
python-dotenv>=1.0.0
google-genai
imagehash
Pillow
opencv-python-headless
numpy
